+++ src/styles/PetStatusAndAlarm.css
