+++ src/styles/SearchBar.css
