+++ src/styles/Timer.css
